tts/pyproject.toml
[project]
name = "tts-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "chatterbox-tts>=0.1.5",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "ml-dtypes==0.5.0",
    "pyaudio>=0.2.14",
    "sounddevice>=0.5.3",
    "torch==2.6.0+cu126",
    "torchaudio==2.6.0+cu126",
]


[tool.uv.sources]
torch = [
  { index = "pytorch-cu126" },
]
torchaudio = [
  { index = "pytorch-cu126" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
